# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SUNAT - Sistema de Facturación Electrónica",
        "description": "API REST para emisión de comprobantes electrónicos SUNAT (Facturas, Boletas, Notas de Crédito/Débito)",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@tuempresa.com"
        },
        "license": {
            "name": "Privada",
            "url": "https://tuempresa.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local Docker"
        }
    ],
    "paths": {
...
        "/api/v1/branches": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Listar sucursales",
                "description": "Obtiene el listado de sucursales, con opción de filtrar por empresa",
                "operationId": "3db09dc381c1a9169fad03df73dc5cab",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Filtrar por ID de empresa",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de sucursales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "company_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "codigo": {
                                                        "type": "string",
                                                        "example": "0000"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Sucursal Principal"
                                                    },
                                                    "direccion": {
                                                        "type": "string",
                                                        "example": "Av. Principal 123"
                                                    },
                                                    "ubigeo": {
                                                        "type": "string",
                                                        "example": "150101"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Crear sucursal",
                "description": "Crea una nueva sucursal o establecimiento anexo",
                "operationId": "b51f2f2107e0c7ebde909393cd27c276",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "codigo",
                                    "nombre",
                                    "direccion",
                                    "ubigeo"
                                ],
                                "properties": {
                                    "company_id": {
                                        "description": "ID de la empresa",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "codigo": {
                                        "description": "Código de establecimiento SUNAT (0000=principal)",
                                        "type": "string",
                                        "example": "0000"
                                    },
                                    "nombre": {
                                        "description": "Nombre de la sucursal",
                                        "type": "string",
                                        "example": "Sucursal Principal"
                                    },
                                    "direccion": {
                                        "description": "Dirección fiscal",
                                        "type": "string",
                                        "example": "Av. Principal 123"
                                    },
                                    "ubigeo": {
                                        "description": "Código de ubigeo (6 dígitos)",
                                        "type": "string",
                                        "example": "150101"
                                    },
                                    "distrito": {
                                        "description": "Distrito",
                                        "type": "string",
                                        "example": "Lima"
                                    },
                                    "provincia": {
                                        "description": "Provincia",
                                        "type": "string",
                                        "example": "Lima"
                                    },
                                    "departamento": {
                                        "description": "Departamento",
                                        "type": "string",
                                        "example": "Lima"
                                    },
                                    "telefono": {
                                        "description": "Teléfono",
                                        "type": "string",
                                        "example": "014567890"
                                    },
                                    "email": {
                                        "description": "Correo electrónico",
                                        "type": "string",
                                        "example": "sucursal@empresa.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucursal creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sucursal creada exitosamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese su token de acceso obtenido del endpoint /api/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}